requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
mysql-connector-python>=8.0.0
flask>=2.3.0
flask-socketio>=5.3.0
Pillow>=9.0.0
psutil>=5.9.0
